rename main.py as spotmodel.py
