+++ excalibur/transit/spotmodel/spotmodel.py
